rule-engine/requirements.txt
fastapi
uvicorn
pika
pymongo
python-logstash~=0.4.8